test: update verify to use correct method
There was an execution difference at:
https://github.com/Bitcoin-ABC/bitcoin-abc/blob/master/src/script/bitcoinconsensus.cpp#L120

There was a new method to the bindings added that includes a new function
signature with an amount argument, that function should be used in the case that
SCRIPT_ENABLE_SIGHASH_FORKID or SCRIPT_FLAGS_VERIFY_WITNESS_DEPRECATED is used.
There was a problem with the test script not runnning the correct methods.